requirements.txt
click
Flask
Flask-HTTPAuth
Flask-SQLAlchemy
itsdangerous
Jinja2
MarkupSafe
PyJWT
SQLAlchemy
Werkzeug
Flask-RestPlus